requirements.txt
flask
face_recognition
numpy
opencv-python
gunicorn
python-dotenv